Add support for maintaining firmware built from source files
This commit adds support for keeping firmware source in the debian
source package. This allows the firmware to be rebuilt by the package
maintainer as needed. The fw_source directory is not included in the
resulting package.

To build the firmware from source, enter the fw_source directory and
execute 'make'. To update the firmware in the firmware tree, run 'make
update', which will build the source, install it in the firmware tree,
and clean the built files in fw_source.

Firmware built from source will need to conform to the following:

1. Add the source to a new subdirectory under fw_source
2. Create a Makefile with the following rules:
  - all: build the firmware
  - install: install firmware in firmware tree ($firmware_dir points to
    the firmware tree)
  - clean: clean the built files created by the all phase
3. Add the subdirectory to the SUBDIRS var in fw_source/Makefile

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>